Created components to add, update and remove a bootcamp, added search functionality to Landing component, also added virtual fields to the user model

+++ client/src/redux/actions/bootcamps.js
@@ -5,33 +5,30 @@ import {
   ADD_BOOTCAMP,
   REMOVE_BOOTCAMP,
   BOOTCAMP_ERROR,
+  IS_LOADING,
 } from "./types";
 
 import composeUrl from "../../utils/composeUrl";
+import { createAlert } from "./alert";
 
 // Fetch all bootcamps
-export const getBootcamps = (filters) => async (dispatch) => {
-  const url = composeUrl("/api/v1/bootcamps", filters);
-
+export const getBootcamps = (filters, location) => async (dispatch) => {
+  let url = "/api/v1/bootcamps?limit=4";
+  if (filters) {
+    // Get bootcamps with filters if provided
+    url = composeUrl("/api/v1/bootcamps", filters);
+  } else if (location) {
+    // Get bootcamps by radius
+    const { zipcode, miles } = location;
+    // Check if both zipcode and miles are not empty
+    if (zipcode && miles) {
+      url = `${url}/radius/${zipcode}/${miles}`;
+    }
+  }
+  console.log(location);
+  console.log(url);
   try {
     const res = await axios.get(url);
-
-    dispatch({
-      type: GET_BOOTCAMPS,
-      payload: res.data,
-    });
-  } catch (error) {
-    dispatch({
-      type: BOOTCAMP_ERROR,
-      payload: error,
-    });
-  }
-};
-
-// Get bootcamps by radius
-export const getBootcampsByRadius = (zipcode, miles) => async (dispatch) => {
-  try {
-    const res = await axios.get(`/api/v1/bootcamps/radius/${zipcode}/${miles}`);
 
     dispatch({
       type: GET_BOOTCAMPS,
@@ -63,7 +60,7 @@ export const getBootcamp = (bootcampId) => async (dispatch) => {
 };
 
 //   Add bootcamp
-export const addBootcamp = (formData) => async (dispatch) => {
+export const addBootcamp = (formData, history) => async (dispatch) => {
   try {
     const config = {
       header: {
@@ -73,9 +70,15 @@ export const addBootcamp = (formData) => async (dispatch) => {
     const res = await axios.post("/api/v1/bootcamps", formData, config);
 
     dispatch({
+      type: IS_LOADING,
+    });
+
+    dispatch({
       type: ADD_BOOTCAMP,
       payload: res.data,
     });
+
+    history.push("/manage-bootcamp");
   } catch (error) {
     dispatch({
       type: BOOTCAMP_ERROR,
@@ -100,9 +103,11 @@ export const updateBootcamp = (bootcampId, formData) => async (dispatch) => {
     );
 
     dispatch({
-      type: GET_BOOTCAMPS,
+      type: ADD_BOOTCAMP,
       payload: res.data,
     });
+
+    dispatch(createAlert("Bootcamp updated.", "success", 3000));
   } catch (error) {
     dispatch({
       type: BOOTCAMP_ERROR,
@@ -113,17 +118,26 @@ export const updateBootcamp = (bootcampId, formData) => async (dispatch) => {
 
 // Delete bootcamp
 export const deleteBootcamp = (bootcampId) => async (dispatch) => {
-  try {
-    await axios.delete(`/api/v1/bootcamps/${bootcampId}`);
+  if (window.confirm("Are you sure? This action can NOT be undone.")) {
+    try {
+      await axios.delete(`/api/v1/bootcamps/${bootcampId}`);
 
-    dispatch({
-      type: REMOVE_BOOTCAMP,
-      payload: bootcampId,
-    });
-  } catch (error) {
-    dispatch({
-      type: BOOTCAMP_ERROR,
-      payload: error,
-    });
+      dispatch({
+        type: REMOVE_BOOTCAMP,
+        payload: bootcampId,
+      });
+    } catch (error) {
+      dispatch({
+        type: BOOTCAMP_ERROR,
+        payload: error,
+      });
+    }
   }
+};
+
+// Make the page load after a request is sent
+export const isLoading = () => (dispatch) => {
+  dispatch({
+    type: IS_LOADING,
+  });
 };
